+++ src/Modules/Tokens.js
@@ -1,9 +1,11 @@
 const { ACCESS_TOKEN_EXPIRE, ACCESS_TOKEN_SECRET, REFRESH_TOKEN_EXPIRE, REFRESH_TOKEN_SECRET, EMAIL_TOKEN_SECRET, EMAIL_TOKEN_EXPIRE, MAX_REFRESH_TOKENS, RESET_PASSWORD_SECRET, RESET_PASSWORD_EXPIRE, TEMP_ACCESS_SECRET, TEMP_ACCESS_EXPIRE, CLIENT_URL } = require("./../config");
 
+const { GetStoredRefreshTokens, AddStoredRefreshTokens, RevokeStoredRefreshToken, RevokeAllStoredRefreshTokens} = require("./Database");
 const { randomBytes, randomUUID } = require("crypto");
 const jwt = require("jsonwebtoken");
 const redis = require("redis");
 
+const REDIS_REFRESH_TOKEN_FIELD = "refresh_tokens";
 
 const RedisClient = redis.createClient();
 
@@ -12,12 +14,12 @@ async function CheckRedisConnection(){
 		return RedisClient.connect();
 	}
 }
-
-async function GenerateAccessToken(userid, ip) {
+async function GenerateAccessToken(userid, Role,  ip) {
 	let token = jwt.sign(
 		{
 			UserId: userid,
 			CreatedByIp: ip,
+			Role: Role,
 			CreatedAt: Date.now(),
 			Expires: ACCESS_TOKEN_EXPIRE,
 		},
@@ -61,43 +63,71 @@ async function GeneratePasswordResetToken(email, userid) {
 	return `${CLIENT_URL}/login/password-reset/${token}`;
 }
 
-async function GetRefreshTokens(userid) {
+async function GetRefreshTokens(AccountId){
 	await CheckRedisConnection();
-	let tokens = await RedisClient.hGet("refresh_tokens", userid);
-	if (tokens) {
-		return JSON.parse(tokens);
+	let RedisTokens = await RedisClient.hGet(REDIS_REFRESH_TOKEN_FIELD, AccountId);
+	if (RedisTokens && JSON.parse(RedisTokens).length > 0){
+		return JSON.parse(RedisTokens);
 	}
-	return {};
+	let StoredTokens = await GetStoredRefreshTokens(AccountId);
+	await RedisClient.hSet(REDIS_REFRESH_TOKEN_FIELD, AccountId, JSON.stringify(StoredTokens));
+	return StoredTokens;
 }
 
-async function GetOldestToken(tokens) {
-	let OldestTime = [Date.now(), null];
-	for (let id in tokens) {
-		let CreationTime = tokens[id].CreatedAt;
-		if (CreationTime <= OldestTime[0]) {
-			OldestTime = [CreationTime, id];
+async function RevokeOldestRefreshToken(AccountId, Tokens){
+	let Keys = Object.keys(Tokens);
+	let Oldest = Date.now();
+	let TokenId = null;
+	for (let i = 0; i < Keys.length; i++){
+		let Key = Keys[i];
+		let Token = Tokens[Key];
+		let CreationTime = Number(Token.CreatedAt);
+		if (CreationTime < Oldest){
+			Oldest = CreationTime;
+			TokenId = Key;
 		}
 	}
-	let id = (OldestTime[1]);
-	return id;
+	delete(Tokens[TokenId]);
+	await RedisClient.hDel(REDIS_REFRESH_TOKEN_FIELD, AccountId);
+	await RedisClient.hSet(REDIS_REFRESH_TOKEN_FIELD, AccountId, JSON.stringify(Tokens), async (err) => {
+		console.error(err);
+	});
+	await RevokeStoredRefreshToken(AccountId, TokenId);
+	console.log(`Revoked token with id ${TokenId}`);
+	return Tokens;
 }
 
-async function GenerateRefreshToken(userid, ip) {
+async function RevokeAllRefreshTokens(AccountId){
+	await RedisClient.hDel(REDIS_REFRESH_TOKEN_FIELD, AccountId);
+	await RevokeAllStoredRefreshTokens(AccountId);
+	return;
+}
+
+async function UpdateRefreshTokens(AccountId, Tokens, TokenId, Token, TimeStamp){
 	await CheckRedisConnection();
-	let tokens = await GetRefreshTokens(userid);
-	if (Object.keys(tokens).length >= MAX_REFRESH_TOKENS) {
-		delete (tokens[await GetOldestToken(tokens)]);
+	await RedisClient.hSet(REDIS_REFRESH_TOKEN_FIELD, AccountId, JSON.stringify(Tokens), async (err) => {
+		console.error(err);
+	});
+	await AddStoredRefreshTokens(AccountId, TokenId, Token, TimeStamp);
+	console.warn(`Current records: ${Object.keys(Tokens).length}`);
+	return;
+}
+
+async function GenerateRefreshToken(AccountId, ip) {
+	let Tokens = await GetRefreshTokens(AccountId);
+	if (Object.keys(Tokens).length == MAX_REFRESH_TOKENS){
+		Tokens = RevokeOldestRefreshToken(AccountId, Tokens);
 	}
 	let TokenId = randomUUID();
-
-	let token = jwt.sign(
+	let TimeStamp = Date.now().toString();
+	let Token = jwt.sign(
 		{
 			_id: TokenId,
-			UserId: userid,
-			Value: randomBytes(64).toString("hex"),
+			UserId: AccountId,
+			Gibberish: randomBytes(64).toString("hex"),
 			Expires: REFRESH_TOKEN_EXPIRE,
 			CreatedByIp: ip,
-			CreatedAt: Date.now(),
+			CreatedAt: TimeStamp,
 		},
 		REFRESH_TOKEN_SECRET,
 		{
@@ -105,13 +135,12 @@ async function GenerateRefreshToken(userid, ip) {
 			expiresIn: REFRESH_TOKEN_EXPIRE,
 		},
 	);
-	tokens[TokenId] = {
-		Token: token,
-		CreatedAt: Date.now(),
-		Ip: ip,
+	Tokens[TokenId] = {
+		Token: Token,
+		CreatedAt: TimeStamp,
 	};
-	await RedisClient.hSet("refresh_tokens", userid, JSON.stringify(tokens));
-	return token;
+	await UpdateRefreshTokens(AccountId, Tokens, TokenId, Token, TimeStamp);
+	return Token;
 }
 
 async function GenerateTempAccessToken(Email){
@@ -129,22 +158,6 @@ async function GenerateTempAccessToken(Email){
 	return token;
 }
 
-async function RevokeRefreshToken(UserId, uuid) {
-	await CheckRedisConnection();
-	let tokens = await GetRefreshTokens(UserId);
-	if (tokens[uuid]) {
-		delete (tokens[uuid]);
-		await RedisClient.hSet("refresh_tokens", UserId, JSON.stringify(tokens));
-		return { status: true, message: "Token revoked" };
-	}
-	return { staus: false, message: "token not found" };
-}
-
-async function RefreshAllRefreshTokens(UserId){
-	await CheckRedisConnection();
-	await RedisClient.hDel("refresh_tokens", UserId);
-}
-
 
 module.exports = {
 	GenerateAccessToken: GenerateAccessToken,
@@ -152,6 +165,4 @@ module.exports = {
 	GenerateRefreshToken: GenerateRefreshToken,
 	GeneratePasswordResetToken: GeneratePasswordResetToken,
 	GenerateTempAccessToken: GenerateTempAccessToken,
-	RevokeRefreshToken: RevokeRefreshToken,
-	RefreshAllRefreshTokens: RefreshAllRefreshTokens,
 };